Base colour palettes for IDEM, IDDU, and ACEFA

Base colour palettes for IDEM, IDDU, and ACEFA

Usage

idpal(p = c("idem", "iddu", "acefa", "acefa_pink", "idem_official"))

Arguments

p

character. Which palette do you want?

Value

character of hex values

Details

"idem", "iddu", and "acefa" return palettes for all colours in logos. "..._official" palettes return only the four-colour palettes per the style- guide from the designer.

NB: These 'non-official' colours are estimated by a colour dropper thing Gerry found on the internet and may or may not be exacty correct.

Examples

idpal("idem")
#> [1] "#262261" "#1d4d79" "#178793" "#00a79e" "#58b4ac" "#acd3d0" "#e6e7e8"